+++ pkg/reconciler/ingress/resources/names/doc.go
