+++ pkg/reconciler/ingress/resources/names/names.go
